// web/src/api/models/chat-message.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Feedback } from './feedback';
// May contain unused imports in some cases
// @ts-ignore
import type { Reference } from './reference';

/**
 * 
 * @export
 * @interface ChatMessage
 */
export interface ChatMessage {
    /**
     * 
     * @type {string}
     * @memberof ChatMessage
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatMessage
     */
    'part_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatMessage
     */
    'type'?: ChatMessageTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof ChatMessage
     */
    'timestamp'?: number;
    /**
     * 
     * @type {string}
     * @memberof ChatMessage
     */
    'role'?: ChatMessageRoleEnum;
    /**
     * 
     * @type {string}
     * @memberof ChatMessage
     */
    'data'?: string;
    /**
     * 
     * @type {Array<Reference>}
     * @memberof ChatMessage
     */
    'references'?: Array<Reference>;
    /**
     * 
     * @type {Array<string>}
     * @memberof ChatMessage
     */
    'urls'?: Array<string>;
    /**
     * 
     * @type {Feedback}
     * @memberof ChatMessage
     */
    'feedback'?: Feedback;
    /**
     * 
     * @type {Array<any>}
     * @memberof ChatMessage
     */
    'files'?: Array<any>;
}

export const ChatMessageTypeEnum = {
    welcome: 'welcome',
    message: 'message',
    start: 'start',
    stop: 'stop',
    error: 'error',
    tool_call_result: 'tool_call_result',
    thinking: 'thinking',
    references: 'references'
} as const;

export type ChatMessageTypeEnum = typeof ChatMessageTypeEnum[keyof typeof ChatMessageTypeEnum];
export const ChatMessageRoleEnum = {
    human: 'human',
    ai: 'ai'
} as const;

export type ChatMessageRoleEnum = typeof ChatMessageRoleEnum[keyof typeof ChatMessageRoleEnum];